+++ R/ipf.R
@@ -14,6 +14,9 @@
 #' \code{group} categories is different in \code{source} and \code{target}, the data frame returns
 #' the combination of \code{unit} and \code{group} categories that occur in both datasets.
 #' Zero values are replaced by a small, non-zero number (1e-4).
+#' Note that the values returned sum to the observations of the source data frame, not the
+#' target data frame. This is different from other IPF implementations, but ensures that the IPF
+#' does not change the number of observations.
 #'
 #' @param source A "source" data frame. The marginals of this
 #'   dataset are adjusted to the marginals of \code{target}.
@@ -82,7 +85,7 @@ ipf <- function(source, target, group, unit, weight = NULL,
 }
 
 #' @import data.table
-create_common_data <- function(d1, d2, group, unit, fill_na = 1e-10) {
+create_common_data <- function(d1, d2, group, unit, suppress_warnings = FALSE, fill_na = 1e-10) {
     # generate the crossproduct of common groups and units to
     # preserve all possible combinations
     common_group <- fintersect(d1[, group, with = FALSE], d2[, group, with = FALSE])
@@ -91,6 +94,36 @@ create_common_data <- function(d1, d2, group, unit, fill_na = 1e-10) {
     common_unit$key <- 1
     common <- merge(common_unit, common_group, allow.cartesian = TRUE)
     common[, "key" := NULL]
+
+    # this is optional because mutual_difference handles this internally -- no need
+    # to print a warning here
+    if (suppress_warnings == FALSE) {
+        group_removed_d1 <- fsetdiff(d1[, group, with = FALSE], common_group[, group, with = FALSE])
+        group_removed_d2 <- fsetdiff(d2[, group, with = FALSE], common_group[, group, with = FALSE])
+        unit_removed_d1 <- fsetdiff(d1[, unit, with = FALSE], common_unit[, unit, with = FALSE])
+        unit_removed_d2 <- fsetdiff(d2[, unit, with = FALSE], common_unit[, unit, with = FALSE])
+
+        if (nrow(group_removed_d1) > 0) {
+            warning(paste0(
+                "IPF procedure removed ", nrow(group_removed_d1), " group categories from source, ",
+                "this likely reduced the sample size (check columns n_source and n_target)"))
+        }
+        if (nrow(group_removed_d2) > 0) {
+            warning(paste0(
+                "IPF procedure removed ", nrow(group_removed_d2), " group categories from target, ",
+                "this likely reduced the sample size (check columns n_source and n_target)"))
+        }
+        if (nrow(unit_removed_d1) > 0) {
+            warning(paste0(
+                "IPF procedure removed ", nrow(unit_removed_d1), " units from source, ",
+                "this likely reduced the sample size (check columns n_source and n_target)"))
+        }
+        if (nrow(unit_removed_d1) > 0) {
+            warning(paste0(
+                "IPF procedure removed ", nrow(unit_removed_d2), " units from target, ",
+                "this likely reduced the sample size (check columns n_source and n_target)"))
+        }
+    }
 
     # make sure that fill_na is much smaller than the smallest frequency
     min_freq <- min(c(d1[["freq"]], d2[["freq"]]))
@@ -117,51 +150,45 @@ ipf_compute <- function(data, group, unit,
                         max_iterations = 100, precision = .001,
                         only_group = FALSE, only_unit = FALSE) {
 
+    # work with relative weights
+    data[, p1 := freq1 / sum(freq1)]
+    data[, p2 := freq2 / sum(freq2)]
+    data[, p_group_s := sum(p1), by = group]
+    data[, p_unit_s := sum(p2), by = unit]
+
     if (only_group == TRUE) {
-        # work with relative weights, otherwise will have different counts
-        data[, freq1 := freq1 / sum(freq1)]
-        data[, freq2 := freq2 / sum(freq2)]
-        data[, n_group_t := sum(freq2), by = group]
-        data[, n_unit_t := sum(freq1), by = unit]
+        data[, p_group_t := sum(p2), by = group]
+        data[, p_unit_t := sum(p1), by = unit]
     } else if (only_unit == TRUE) {
-        # work with relative weights, otherwise will have different counts
-        data[, freq1 := freq1 / sum(freq1)]
-        data[, freq2 := freq2 / sum(freq2)]
-        data[, n_group_t := sum(freq1), by = group]
-        data[, n_unit_t := sum(freq2), by = unit]
+        data[, p_group_t := sum(p1), by = group]
+        data[, p_unit_t := sum(p2), by = unit]
     } else {
-        data[, `:=`(n_group_t = sum(freq2)), by = group]
-        data[, `:=`(n_unit_t = sum(freq2)), by = unit]
+        data[, p_group_t := sum(p2), by = group]
+        data[, p_unit_t := sum(p2), by = unit]
     }
 
-    data[, `:=`(n_group_s = sum(freq1)), by = group]
-    data[, `:=`(n_unit_s = sum(freq1)), by = unit]
-    data$n_source <- data$freq1
-
     # IPF algorithm
-    precision <- abs(log(1 / (1 + precision)))
     converged <- FALSE
-    start_i <- round(stats::runif(1, 1, 2)) # start randomly with either row or column adjustment
+    start_i <- sample(2, 1) # start randomly with either row or column adjustment
     max_iterations <- max_iterations - 1 + start_i # offset so max_iterations is still the same
 
     for (i in start_i:(max_iterations * 2)) {
         if (i %% 2 == 0) {
-            data[, `:=`(freq1 = freq1 * n_unit_t / n_unit_s)]
+            data[, `:=`(p1 = p1 * p_unit_t / p_unit_s)]
         } else {
-            data[, `:=`(freq1 = freq1 * n_group_t / n_group_s)]
+            data[, `:=`(p1 = p1 * p_group_t / p_group_s)]
         }
+        data[, p1 := p1 / sum(p1)]
 
         if (i %% 5 == 0) {
             update_log_progress("#")
         }
 
-        data[, `:=`(n_group_s = sum(freq1)), by = group]
-        data[, `:=`(n_unit_s = sum(freq1)), by = unit]
+        data[, p_group_s := sum(p1), by = group]
+        data[, p_unit_s := sum(p1), by = unit]
 
-        group_ratio <- data[, list(first(n_group_s), first(n_group_t)), by = group][,
-            abs(log(V1 / V2))]
-        unit_ratio <- data[, list(first(n_unit_s), first(n_unit_t)), by = unit][,
-            abs(log(V1 / V2))]
+        group_ratio <- data[, list(first(p_group_s), first(p_group_t)), by = group][, abs(V1 - V2)]
+        unit_ratio <- data[, list(first(p_unit_s), first(p_unit_t)), by = unit][, abs(V1 - V2)]
 
         if (all(group_ratio <= precision) & all(unit_ratio <= precision)) {
             converged <- TRUE
@@ -173,10 +200,14 @@ ipf_compute <- function(data, group, unit,
         close_log()
         stop("IPF did not converge; lower precision or increase max_iterations.")
     }
-    data[, c("n_unit_t", "n_unit_s", "n_group_t", "n_group_s",
-             "freq_orig1", "freq_orig2") := NULL]
-    setnames(data, "freq1", "n")
+
+    # adjust ipf-adjusted count to sample size of source
+    data[, n := p1 / sum(p1) * sum(freq1)]
+    setnames(data, "freq1", "n_source")
     setnames(data, "freq2", "n_target")
+
+    data[, c("p_unit_t", "p_unit_s", "p_group_t", "p_group_s",
+             "freq_orig1", "freq_orig2", "p1", "p2") := NULL]
     setkey(data, NULL)
     data
 }