+++ R/segregation.R
@@ -12,14 +12,12 @@
 NULL
 
 globalVariables(c(
-    "cond1", "cond2", "entropy_cond", "entropy_cond1", "entropy_cond2", "entropyw",
-    "est", "freq", "freq1", "freq2", "freq_orig1", "freq_orig2",
-    "ls_unit", "n_group", "n_group_s", "n_group_t",
-    "n_group_target", "n_source", "n_target", "n_unit", "n_unit_s", "n_unit_t",
-    "n_unit_target", "n_within_group", "p", "p_group", "p_group_g_unit", "p_group_g_unit1",
-    "p_group_g_unit2", "p_unit", "p_unit1", "p_unit2", "p_within", "ratio", "sumcond1",
-    "sumcond2", "unit1", "unit2",
-    "ls_diff1", "ls_diff2", "p1", "p2", "total", "V1", "V2"))
+    "V1", "V2", "cond1", "cond2", "entropy_cond", "entropy_cond1", "entropy_cond2", "entropyw",
+    "est", "freq", "freq1", "freq2", "freq_orig1", "freq_orig2", "ls_diff1", "ls_diff2", "ls_unit",
+    "n", "n_group", "n_group_target", "n_source", "n_target", "n_unit", "n_unit_target",
+    "n_within_group", "p", "p1", "p2", "p_group", "p_group_g_unit", "p_group_g_unit1",
+    "p_group_g_unit2", "p_group_s", "p_group_t", "p_unit", "p_unit1", "p_unit2", "p_unit_s",
+    "p_unit_t", "p_within", "sumcond1", "sumcond2", "total", "unit1", "unit2"))
 
 # log